Run analysis on startup and format output.
The output now looks as described in README.md.
On startup, an analysis is performed.
Files which NO LONGER contain violations are now
reported with an empty block of
BEGIN-ANALYSIS xxx
END-ANALYSIS xxx
to allow VSC recognizing that the file is now
"clean" from PMD's point of view.
Besides, a little code refactoring.